pyproject.toml
[project]
name = "django-task"
version = "0.1.0"
description = ""
authors = [
    {name = "Volodymyr Popov",email = "smartappsvp@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "weasyprint (>=65.0,<66.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "django-task"
package-mode = false
